requirements.txt
fastapi
pillow
torch
torchvision
uvicorn
easyocr
opencv-python-headless
